Refactor GraphQL query logging to be much more helpful.
Extract the query execution ID and add it as an MDC variable.
Updated the default log formatter to include the `graphql-query` MDC field.

Added additional logging interceptor to parse the input Query/Mutation and extract the requested fields, along with the provided arguments.
