+++ backend/src/main/java/gov/cdc/usds/simplereport/logging/GraphQLLoggingHelpers.java
@@ -1,11 +1,17 @@
 package gov.cdc.usds.simplereport.logging;
 
+import graphql.ExecutionResult;
+import graphql.execution.instrumentation.InstrumentationContext;
+import graphql.execution.instrumentation.SimpleInstrumentationContext;
 import graphql.language.Field;
 import graphql.language.Selection;
 import graphql.language.SelectionSet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.slf4j.MDC;
 
+import java.time.Duration;
+import java.time.temporal.ChronoUnit;
 import java.util.stream.Stream;
 
 /**
@@ -14,6 +20,7 @@ import java.util.stream.Stream;
 public class GraphQLLoggingHelpers {
 
     private static final Logger LOG = LoggerFactory.getLogger(GraphQLLoggingHelpers.class);
+    public static final String GRAPHQL_QUERY_MDC_KEY = "graphql-query";
 
     private GraphQLLoggingHelpers() {
         // Not used
@@ -30,7 +37,7 @@ public class GraphQLLoggingHelpers {
      * @param selection  - {@link Selection} to process
      * @return - {@link Stream} of field name {@link String} values
      */
-    public static Stream<String> walkFields(String parentName, Selection selection) {
+    public static Stream<String> walkFields(String parentName, Selection<?> selection) {
         final Field field = (Field) selection;
         final String fieldName = field.getName();
 
@@ -50,5 +57,33 @@ public class GraphQLLoggingHelpers {
         }
 
         return selections.getSelections().stream().filter(s -> s instanceof Field).flatMap(f -> GraphQLLoggingHelpers.walkFields(fieldName, f));
+    }
+
+    /**
+     * Create {@link InstrumentationContext} to handle cleanup and logging actions when the associated Query completes
+     * Handles logging the exeuction time, query success/fail and any associated errors
+     *
+     * @param queryStart - {@link Long} timestamp of when query started
+     * @return - {@link InstrumentationContext} to handle query completion
+     */
+    public static InstrumentationContext<ExecutionResult> createInstrumentationContext(long queryStart) {
+        return new SimpleInstrumentationContext<>() {
+            @Override
+            public void onCompleted(ExecutionResult result, Throwable t) {
+                final long queryEnd = System.currentTimeMillis();
+                final Duration queryDuration = Duration.of(queryEnd - queryStart, ChronoUnit.MILLIS);
+
+                if (t != null) {
+                    LOG.error("GraphQL execution failed: {}", t.getMessage(), t);
+                } else if (!result.getErrors().isEmpty()) {
+                    result.getErrors().forEach(error -> LOG.error("Query failed with error {}", error));
+                    LOG.info("Graphql execution failed in {}ms", queryDuration.toMillis());
+                } else {
+                    LOG.info("GraphQL execution completed in {}ms", queryDuration.toMillis());
+                }
+                // Clear the MDC context
+                MDC.remove(GRAPHQL_QUERY_MDC_KEY);
+            }
+        };
     }
 }
